IS - 307
git-svn-id: https://svn.omertex.com:8444/svn/iSimple@487 d8ebeebf-d64f-4e48-9a47-8608552ece91

+++ sources/isimple-project/src/com/treelev/isimple/domain/db/DeliveryZone.java
@@ -77,7 +77,7 @@ public class DeliveryZone {
         this.address = address;
     }
 
-    public double getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
